move mz_repr::statement_logging module to its own file
